+++ BookBaram/View/Edit/EditView.swift
@@ -1,0 +1,75 @@
+//
+//  EditView.swift
+//  BookBaram
+//
+//  Created by 이송미 on 6/17/24.
+//
+
+import UIKit
+
+class EditView: UIView {
+    let scrollView: UIScrollView = {
+        let scrollView = UIScrollView()
+        scrollView.translatesAutoresizingMaskIntoConstraints = false
+        return scrollView
+    }()
+    
+    let contentView: UIStackView = {
+        let stackView = UIStackView()
+        stackView.translatesAutoresizingMaskIntoConstraints = false
+        stackView.axis = .vertical
+        stackView.alignment = .fill
+        stackView.distribution = .fill
+        stackView.spacing = 10.0
+        return stackView
+    }()
+    
+    let imageView: UIImageView = {
+        let imageView = UIImageView()
+        imageView.translatesAutoresizingMaskIntoConstraints = false
+        imageView.backgroundColor = .black // placeholder
+        return imageView
+    }()
+    
+    let titleTextField: UITextField = {
+        let textField = UITextField()
+        textField.translatesAutoresizingMaskIntoConstraints = false
+        return textField
+    }()
+    
+    let contentTextView: UITextView = {
+        let textView = UITextView()
+        textView.translatesAutoresizingMaskIntoConstraints = false
+        return textView
+    }()
+    
+    func layout() {
+        addSubview(scrollView)        
+        addSubview(contentView)
+        contentView.addArrangedSubview(imageView)
+        contentView.addArrangedSubview(titleTextField)
+        contentView.addArrangedSubview(contentTextView)
+        
+        let marginConstant = 15.0
+        
+        scrollView.widthAnchor.constraint(equalTo: widthAnchor).isActive = true
+        scrollView.topAnchor.constraint(equalTo: safeAreaLayoutGuide.topAnchor).isActive = true
+        scrollView.bottomAnchor.constraint(equalTo: safeAreaLayoutGuide.bottomAnchor).isActive = true
+        
+        contentView.leadingAnchor.constraint(equalTo: scrollView.leadingAnchor, constant: marginConstant).isActive = true
+        contentView.trailingAnchor.constraint(equalTo: scrollView.trailingAnchor, constant: marginConstant * -1).isActive = true
+        contentView.widthAnchor.constraint(equalTo: scrollView.widthAnchor, constant: marginConstant * -2).isActive = true
+        contentView.heightAnchor.constraint(equalTo: scrollView.heightAnchor).isActive = true
+
+        imageView.heightAnchor.constraint(equalToConstant: 100.0).isActive = true
+        
+        titleTextField.borderStyle = .roundedRect
+        titleTextField.placeholder = "Write the title"
+        
+        contentTextView.layer.borderWidth = 1.0
+        contentTextView.layer.cornerRadius = 8
+        contentTextView.layer.borderColor = UIColor.lightGray.cgColor
+        contentTextView.text = "Content"
+    }
+
+}
