+++ BookBaram/View/Edit/EditViewController.swift
@@ -8,40 +8,7 @@
 import UIKit
 
 class EditViewController: UIViewController {
-    lazy var scrollView: UIScrollView = {
-        let scrollView = UIScrollView()
-        scrollView.translatesAutoresizingMaskIntoConstraints = false
-        return scrollView
-    }()
-    
-    lazy var contentView: UIStackView = {
-        let stackView = UIStackView()
-        stackView.translatesAutoresizingMaskIntoConstraints = false
-        stackView.axis = .vertical
-        stackView.alignment = .fill
-        stackView.distribution = .fill
-        stackView.spacing = 10.0
-        return stackView
-    }()
-    
-    lazy var imageView: UIImageView = {
-        let imageView = UIImageView()
-        imageView.translatesAutoresizingMaskIntoConstraints = false
-        imageView.backgroundColor = .black // placeholder
-        return imageView
-    }()
-    
-    lazy var titleTextField: UITextField = {
-        let textField = UITextField()
-        textField.translatesAutoresizingMaskIntoConstraints = false
-        return textField
-    }()
-    
-    lazy var contentTextView: UITextView = {
-        let textView = UITextView()
-        textView.translatesAutoresizingMaskIntoConstraints = false
-        return textView
-    }()
+    let editView = EditView()
     
     override func viewDidLoad() {
         super.viewDidLoad()
@@ -52,34 +19,14 @@ class EditViewController: UIViewController {
     
     private func setLayout() {
         self.view.backgroundColor = .systemBackground
+        view.addSubview(editView)
         
-        self.view.addSubview(scrollView)
-        scrollView.addSubview(contentView)
+        editView.translatesAutoresizingMaskIntoConstraints = false
+        editView.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor).isActive = true
+        editView.bottomAnchor.constraint(equalTo: view.safeAreaLayoutGuide.bottomAnchor).isActive = true
+        editView.leadingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.leadingAnchor).isActive = true
+        editView.trailingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.trailingAnchor).isActive = true
         
-        contentView.addArrangedSubview(imageView)
-        contentView.addArrangedSubview(titleTextField)
-        contentView.addArrangedSubview(contentTextView)
-        
-        let marginConstant = 15.0
-        
-        scrollView.widthAnchor.constraint(equalTo: self.view.widthAnchor).isActive = true
-        scrollView.topAnchor.constraint(equalTo: self.view.safeAreaLayoutGuide.topAnchor).isActive = true
-        scrollView.bottomAnchor.constraint(equalTo: self.view.safeAreaLayoutGuide.bottomAnchor).isActive = true
-        
-        contentView.leadingAnchor.constraint(equalTo: scrollView.leadingAnchor, constant: marginConstant).isActive = true
-        contentView.trailingAnchor.constraint(equalTo: scrollView.trailingAnchor, constant: marginConstant * -1).isActive = true
-        contentView.widthAnchor.constraint(equalTo: scrollView.widthAnchor, constant: marginConstant * -2).isActive = true
-        contentView.heightAnchor.constraint(equalTo: scrollView.heightAnchor).isActive = true
-
-        imageView.heightAnchor.constraint(equalToConstant: 100.0).isActive = true
-        
-        titleTextField.borderStyle = .roundedRect
-        titleTextField.placeholder = "Write the title"
-        
-        contentTextView.layer.borderWidth = 1.0
-        contentTextView.layer.cornerRadius = 8
-        contentTextView.layer.borderColor = UIColor.lightGray.cgColor
-        contentTextView.text = "Content"
-                
+        editView.layout()
     }
 }
